feat(mastodon): memory the last selected server and the access history
    save the last selected Mastodon server and the access history in the
    local storage, and reload it when launching the app again
