+++ lib/features/extensions.dart
@@ -1,5 +1,6 @@
 // The core exposed extensions for the Glacial app.
 
+export 'mastodon/extensions.dart';
 export 'preference/extensions.dart';
 
 // vim: set ts=2 sw=2 sts=2 et:
