+++ lib/features/mastodon/extensions.dart
@@ -1,0 +1,29 @@
+// The extensions implementation for the glacial feature.
+import 'dart:async';
+import 'dart:convert';
+
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'package:glacial/core.dart';
+import 'package:glacial/features/models.dart';
+
+extension AccessStatusExtension on Storage {
+  // Load the access status from the storage.
+  Future<AccessStatusSchema?> loadAccessStatus({WidgetRef? ref}) async {
+    final String? json = await getString(AccessStatusSchema.key);
+    final AccessStatusSchema? status = json == null ? null : AccessStatusSchema.fromString(json);
+
+    ref?.read(accessStatusProvider.notifier).state = status;
+    return status;
+  }
+
+  // Save the access status to the storage.
+  Future<void> saveAccessStatus(AccessStatusSchema schema, {WidgetRef? ref}) async {
+    final String json = jsonEncode(schema.toJson());
+    await setString(AccessStatusSchema.key, json);
+
+    ref?.read(accessStatusProvider.notifier).state = schema;
+  }
+}
+
+// vim: set ts=2 sw=2 sts=2 et:
